Add count all comment api test
