+++ src/main/resources/data.properties
@@ -1,1 +1,3 @@
 GreencityUserAPI=https://greencity-user.greencity.cx.ua
+GreencityAPI=https://greencity.greencity.cx.ua
+GreencityToken=
